# backend/openapi/horse-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Horse Reservation API",
    "description": "마장 예약 서비스 API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/admin/timeslots/{timeSlotId}/reservations/{reservationId}/cancel-for-closure": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "cancelReservation",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleClosureCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancelResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-dates/{scheduleDate}/reservations/{reservationId}/cancel-for-closure": {
      "post": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "cancelReservation_1",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleClosureCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/schedule-dates/{scheduleDate}/closing": {
      "post": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "startClosing",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDateActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDateClosureImpactResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/schedule-dates/{scheduleDate}/closing/cancel": {
      "post": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "cancelClosing",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDateActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDateClosureImpactResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/schedule-dates/{scheduleDate}/close": {
      "post": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "close",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDateActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDateClosureImpactResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-dates": {
      "get": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "getScheduleDates",
        "parameters": [
          {
            "name": "dateFrom",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleDateResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/schedule-dates/{scheduleDate}": {
      "get": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "getScheduleDate",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/schedule-dates/{scheduleDate}/closure-impact": {
      "get": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "getClosureImpact",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDateClosureImpactResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReservationCancelResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "responsibility": {
            "type": "string"
          },
          "couponAction": {
            "type": "string"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time"
          },
          "changed": {
            "type": "boolean"
          }
        }
      },
...
      "ScheduleClosureCancelRequest": {
        "type": "object",
        "properties": {
          "memo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          }
        },
        "required": [
          "memo"
        ]
      },
...
      "ScheduleDateActionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "expectedVersion",
          "reason"
        ]
      },
      "ScheduleDateClosureImpactResponse": {
        "type": "object",
        "properties": {
          "scheduleDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "NORMAL",
              "OPEN",
              "CLOSING",
              "CLOSED"
            ]
          },
          "resumeStatus": {
            "type": "string",
            "enum": [
              "NORMAL",
              "OPEN",
              "CLOSING",
              "CLOSED"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "initialReservationCount": {
            "type": "integer",
            "format": "int32"
          },
          "activeReservationCount": {
            "type": "integer",
            "format": "int32"
          },
          "resolvedReservationCount": {
            "type": "integer",
            "format": "int32"
          },
          "remainingReservationCount": {
            "type": "integer",
            "format": "int32"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleImpactReservationResponse"
            }
          },
          "changed": {
            "type": "boolean"
          }
        },
        "required": [
          "activeReservationCount",
          "changed",
          "initialReservationCount",
          "progressPercent",
          "remainingReservationCount",
          "reservations",
          "resolvedReservationCount",
          "scheduleDate",
          "status",
          "version"
        ]
      },
      "ScheduleImpactReservationResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "memberName": {
            "type": "string"
          },
          "memberPhone": {
            "type": "string"
          },
          "classType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "paymentSource": {
            "type": "string"
          },
          "couponId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "classType",
          "memberId",
          "memberName",
          "memberPhone",
          "paymentSource",
          "reservationId",
          "status"
        ]
      },
...
      "ScheduleDateResponse": {
        "type": "object",
        "properties": {
          "scheduleDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "NORMAL",
              "OPEN",
              "CLOSING",
              "CLOSED"
            ]
          },
          "resumeStatus": {
            "type": "string",
            "enum": [
              "NORMAL",
              "OPEN",
              "CLOSING",
              "CLOSED"
            ]
          },
          "reason": {
            "type": "string"
          },
          "appliedConfigVersion": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "appliedConfigVersion",
          "scheduleDate",
          "status",
          "version"
        ]
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "details",
          "fieldErrors",
          "message",
          "path",
          "status",
          "timestamp"
        ]
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}